pyproject.toml
[tool.poetry]
name = "gpt-trainer-sdk"
version = "0.1.3"
description = ""
authors = ["Michael T. Chang <michael.chang@petal.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
